Start porting Qt to 64-bit OS X.
This change should fix most of the trivial compiler errors.
QMacSavedPortInfo has been ifdefed out, this class should not
be needed since QuickDraw is no longer supported.

[git-p4: depot-paths = "//depot/qt/main/": change = 239574]

+++ src/gui/painting/qpaintdevice_mac.cpp
@@ -112,7 +112,7 @@ Q_GUI_EXPORT CGContextRef qt_mac_cg_context(const QPaintDevice *pdev)
             return 0;
 
         if(OSStatus err = CreateCGContextForPort(port, &ret)) {
-            qWarning("QPaintDevice: Unable to create CGContext for port %p [%ld]", port, err);
+            qWarning("QPaintDevice: Unable to create CGContext for port %p [%ld]", port, long(err));
             return 0;
         }
         SyncCGContextOriginWithPort(ret, port);
